+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Tests for second purchase without geolocation and postal code.
+
 ## [0.3.1] - 2020-11-23
 
 ### Fixed